+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `Opis\View\EngineEntry` class was moved into its own file
 - `Opis\View\ViewRouter` was renamed to `Opis\View\ViewApp`
 - Source files were moved from `lib` to `src`
+- `Opis\View\ViewableInterface` was renamed to `Opis\View\IView`
 
 ### Fixed
 - Nothing